+++ src/test/MacroReplacement.test.ts
@@ -1,76 +1,53 @@
-import {
-    TextDocument,
-    createConnection,
-    Diagnostic,
-    DiagnosticSeverity,
-    Range,
-    Position,
-    TextDocuments,
-    DidOpenTextDocumentParams,
-    DidOpenTextDocumentNotification,
-    DidCloseTextDocumentParams,
-    DidCloseTextDocumentNotification,
-    TextDocumentSyncKind,
-    ProposedFeatures,
-    ConnectionStrategy
-} from 'vscode-languageserver';
+import { TextDocument, Range, Position } from 'vscode-languageserver';
 import * as vscode from 'vscode';
-import {
-    Uri, workspace
-} from 'vscode';
+import { Uri, workspace } from 'vscode';
 import * as path from 'path';
-import * as fs from 'fs';
 import * as assert from 'assert';
 import { ANTLRBackend } from '../compiling/ANTLRBackend';
-import { getPathFromUri } from '../utils/common';
-import { beforeEach } from 'mocha'
-import { Select_expressionContext } from '../compiling/ANTLR/grammar/build/SystemVerilogParser';
-import { TransportKind, TextDocumentIdentifier } from 'vscode-languageclient';
-import { Duplex } from 'stream';
 
-const testFolderLocation = '../../src/test/';
+const testFolderLocation = '../../src/test';
 
 suite('Macro Replacement Tests', () => {
     test('test #1: Macro replacement/definition removal for text without macro definitions', async () => {
-        let input_file_name: string = `test-files/MacroReplace.test/no_macros.sv`;
-        let expected_file_name: string = `test-files/MacroReplace.test/no_macros_expected.sv`;
+        const input_file_name: string = 'test-files/MacroReplace.test/no_macros.sv';
+        const expected_file_name: string = 'test-files/MacroReplace.test/no_macros_expected.sv';
         await macroReplaceTest(input_file_name, expected_file_name);
     }).timeout(10000);
 
     test('test #2: Macro replacement/definition removal for text with single macro definition but no macro uses', async () => {
-        let input_file_name: string = `test-files/MacroReplace.test/single_macro_no_uses.sv`;
-        let expected_file_name: string = `test-files/MacroReplace.test/single_macro_no_uses_expected.sv`;
+        const input_file_name: string = 'test-files/MacroReplace.test/single_macro_no_uses.sv';
+        const expected_file_name: string = 'test-files/MacroReplace.test/single_macro_no_uses_expected.sv';
         await macroReplaceTest(input_file_name, expected_file_name);
     }).timeout(10000);
 
     test('test #3: Macro replacement/definition removal for text with single macro definition with multiple macro uses', async () => {
-        let input_file_name: string = `test-files/MacroReplace.test/single_macro_multiple_uses.sv`;
-        let expected_file_name: string = `test-files/MacroReplace.test/single_macro_multiple_uses_expected.sv`;
+        const input_file_name: string = 'test-files/MacroReplace.test/single_macro_multiple_uses.sv';
+        const expected_file_name: string = 'test-files/MacroReplace.test/single_macro_multiple_uses_expected.sv';
         await macroReplaceTest(input_file_name, expected_file_name);
     }).timeout(10000);
 
     test('test #4: Macro replacement/definition removal for text with multiple macro definitions with multiple macro uses', async () => {
-        let input_file_name: string = `test-files/MacroReplace.test/multiple_macros_multiple_uses.sv`;
-        let expected_file_name: string = `test-files/MacroReplace.test/multiple_macros_multiple_uses_expected.sv`;
+        const input_file_name: string = 'test-files/MacroReplace.test/multiple_macros_multiple_uses.sv';
+        const expected_file_name: string = 'test-files/MacroReplace.test/multiple_macros_multiple_uses_expected.sv';
         await macroReplaceTest(input_file_name, expected_file_name);
     }).timeout(10000);
 
     test('test #5: Macro replacement/definition removal for text with multiple macro definitions interspersed throughout the file', async () => {
-        let input_file_name: string = `test-files/MacroReplace.test/multiple_macros_interspersed.sv`;
-        let expected_file_name: string = `test-files/MacroReplace.test/multiple_macros_interspersed_expected.sv`;
+        const input_file_name: string = 'test-files/MacroReplace.test/multiple_macros_interspersed.sv';
+        const expected_file_name: string = 'test-files/MacroReplace.test/multiple_macros_interspersed_expected.sv';
         await macroReplaceTest(input_file_name, expected_file_name);
     }).timeout(10000);
 
-    //This feature is not currently implemented
-    /*test('test #6: Macro replacement/definition removal for text with single macro defined twice and used after each definition', async () => {
-        let input_file_name: string = `test-files/MacroReplace.test/redefined_macro.sv`;
-        let expected_file_name: string = `test-files/MacroReplace.test/redefined_macro_expected.sv`;
+    // This feature is not currently implemented
+    /* test('test #6: Macro replacement/definition removal for text with single macro defined twice and used after each definition', async () => {
+        let input_file_name: string = 'test-files/MacroReplace.test/redefined_macro.sv';
+        let expected_file_name: string = 'test-files/MacroReplace.test/redefined_macro_expected.sv';
         await macroReplaceTest(input_file_name, expected_file_name);
-    }).timeout(10000);*/
+    }).timeout(10000); */
 
     test('test #7: Macro replacement/definition removal for text with single multiline macro definition with one use', async () => {
-        let input_file_name: string = `test-files/MacroReplace.test/multiline_macro.sv`;
-        let expected_file_name: string = `test-files/MacroReplace.test/multiline_macro_expected.sv`;
+        const input_file_name: string = 'test-files/MacroReplace.test/multiline_macro.sv';
+        const expected_file_name: string = 'test-files/MacroReplace.test/multiline_macro_expected.sv';
         await macroReplaceTest(input_file_name, expected_file_name);
     }).timeout(10000);
 });
@@ -81,30 +58,30 @@ suite('Macro Replacement Tests', () => {
  * @param expected_file The file to retrieve the expected text from
  */
 async function macroReplaceTest(input_file: string, expected_file: string) {
-        let filePath = path.join(__dirname, testFolderLocation, input_file);
-        let uriDoc = Uri.file(filePath);
+    const filePath = path.join(__dirname, testFolderLocation, input_file);
+    const uriDoc = Uri.file(filePath);
 
-        let documentWorkspace = await workspace.openTextDocument(uriDoc);
-        let document: TextDocument = castTextDocument(documentWorkspace);
-        let input_text: string = document.getText();
+    const documentWorkspace = await workspace.openTextDocument(uriDoc);
+    const document: TextDocument = castTextDocument(documentWorkspace);
+    const input_text: string = document.getText();
 
-        let filePath2 = path.join(__dirname, testFolderLocation, expected_file);
-        let uriDoc2 = Uri.file(filePath2);
+    const filePath2 = path.join(__dirname, testFolderLocation, expected_file);
+    const uriDoc2 = Uri.file(filePath2);
 
-        let documentWorkspace2 = await workspace.openTextDocument(uriDoc2);
-        let document2: TextDocument = castTextDocument(documentWorkspace2);
-        let expected_text: string = document2.getText().replace(/\r\n/g, '\n');
+    const documentWorkspace2 = await workspace.openTextDocument(uriDoc2);
+    const document2: TextDocument = castTextDocument(documentWorkspace2);
+    const expected_text: string = document2.getText().replace(/\r\n/g, '\n');
 
-        let output_text: string = new ANTLRBackend().macroReplace(input_text);
+    const output_text: string = new ANTLRBackend().macroReplace(input_text);
 
-        if (output_text != expected_text) {
-            assert.fail();
-        } // else pass
+    if (output_text !== expected_text) {
+        assert.fail();
+    } // else pass
 }
 
 /**
  * Converts a given `document` from vscode.TextDocument to vscode-languageserver.TextDocument
- * 
+ *
  * @param document the document to convert
  * @returns a converted document
  */
@@ -114,14 +91,12 @@ function castTextDocument(document: vscode.TextDocument): TextDocument {
         languageId: document.languageId,
         version: document.version,
         getText(range?: Range): string {
-            if (range)
-                return document.getText(castRange(range));
-            else
-                return document.getText();
+            if (range) return document.getText(castRange(range));
+            return document.getText();
         },
         lineCount: document.lineCount,
         positionAt(offset: number): Position {
-            let position = document.positionAt(offset);
+            const position = document.positionAt(offset);
             return {
                 line: position.line,
                 character: position.character
@@ -135,22 +110,22 @@ function castTextDocument(document: vscode.TextDocument): TextDocument {
 
 /**
  * Converts a given `range` from vscode-languageserver.Range to vscode.Range
- * 
+ *
  * @param document the range to convert
  * @returns a converted range
  */
 function castRange(range: Range) {
-    let startOld = range.start;
-    let endOld = range.end;
+    const startOld = range.start;
+    const endOld = range.end;
 
-    let start = new vscode.Position(startOld.line, startOld.character);
-    let end = new vscode.Position(endOld.line, endOld.character);
+    const start = new vscode.Position(startOld.line, startOld.character);
+    const end = new vscode.Position(endOld.line, endOld.character);
 
     return new vscode.Range(start, end);
 }
 /**
  * Converts a given `position` from vscode-languageserver.Position to vscode.Position
- * 
+ *
  * @param document the position to convert
  * @returns a converted position
  */